Correctly cast ctypes function input to unsigned int
This fixes issues found by building on NetBSD which is more pedantic
about thoses cases.

Reported by: Thomas Klausner (wiz at NetBSD)

+++ libpkgconf/argvsplit.c
@@ -61,7 +61,7 @@ pkgconf_argv_split(const char *src, int *argc, char ***argv)
 			}
 			*dst_iter++ = *src_iter;
 		}
-		else if (isspace(*src_iter))
+		else if (isspace((unsigned int)*src_iter))
 		{
 			if ((*argv)[argc_count] != NULL)
 			{
